Refactor derive into its own module
Fix trait derivation for structs (proving the need for tests)

Re-add @clux's vec/clone refactor to reduce allocation
