+++ src/lib.rs
@@ -4,3 +4,5 @@ mod analyzer;
 pub use analyzer::{analyze, Config};
 mod output;
 pub use output::{Container, MapType, Member, Output};
+mod derive;
+pub use derive::{Derive, DeriveTarget};
